+++ src/app/core/services/auth/auth.service.ts
@@ -123,6 +123,8 @@ export class AuthService {
                                     // 3. ASK FOR LOGIN:
                                     this.login(state);
                                     return Promise.resolve();
+                                } else {
+                                    this.logout();
                                 }
                                 return Promise.reject(result);
                             });
@@ -232,7 +234,10 @@ export class AuthService {
     }
 
     logout() {
-        this.oauthService.logOut();
+        if (this.oauthService.hasValidIdToken()) {
+            this.oauthService.logOut(true);
+        }
+        this.router.navigateByUrl('/marketplace');
         localStorage.clear();
     }