+++ src/app/modules/marketplace/services/modules-service/modules.service.ts
@@ -1,3 +1,4 @@
+import { module } from './../tools-service/tools.service.mock';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from '@environments/environment';
@@ -5,11 +6,12 @@ import {
     Module,
     ModuleConfiguration,
     ModuleSummary,
+    Service,
 } from '@app/shared/interfaces/module.interface';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { AppConfigService } from '@app/core/services/app-config/app-config.service';
-import { TagObject } from '@app/data/types/tags';
-
+import { Client } from '@grycap/oscar-js';
+import { OAuthStorage } from 'angular-oauth2-oidc';
 const { base, endpoints } = environment.api;
 
 @Injectable({
@@ -18,17 +20,18 @@ const { base, endpoints } = environment.api;
 export class ModulesService {
     constructor(
         private http: HttpClient,
-        private appConfigService: AppConfigService
+        private appConfigService: AppConfigService,
+        private authStorage: OAuthStorage
     ) {}
 
     readonly voParam = new HttpParams().set('vo', this.appConfigService.voName);
 
-    getModulesSummary(tags?: TagObject): Observable<ModuleSummary[]> {
+    getModulesSummary(tags?: any): Observable<ModuleSummary[]> {
         const url = `${base}${endpoints.modulesSummary}`;
         if (tags) {
             let params = new HttpParams();
             Object.keys(tags).forEach((key: string) => {
-                params = params.set(key, tags[key as keyof TagObject] || '');
+                params = params.set(key, tags[key]);
             });
             return this.http.get<Array<ModuleSummary>>(url, { params });
         } else {
@@ -51,5 +54,70 @@ export class ModulesService {
         return this.http.get<ModuleConfiguration>(url, {
             params: this.voParam,
         });
+    }
+
+    getAccessToken() {
+        const oidc_token = this.authStorage.getItem('access_token');
+        if (!oidc_token) {
+            throw new Error('Authorization error. the token cannot be null.');
+        }
+        return oidc_token;
+    }
+
+    //OSCAR cluster
+    /**
+     * Get list of services in OSCAR
+     * @returns Service list
+     */
+    getServices(oscar_endpoint: string): Promise<Service[]> {
+        const oidc_token = this.getAccessToken();
+        const client: Client = new Client({
+            clusterId: '1',
+            oscar_endpoint,
+            oidc_token,
+        });
+        return client.getServices();
+    }
+
+    /**
+     * Run service in OSCAR usign sync inovoation
+     * @param oscar_endpoint oscar endpoint
+     * @param serviceName module name
+     * @param file file to run
+     * @returns response with the result of the service.
+     */
+    runService(
+        oscar_endpoint: string,
+        serviceName: string,
+        file: string
+    ): Promise<any> {
+        const oidc_token = this.getAccessToken();
+        const client: Client = new Client({
+            clusterId: '1',
+            oscar_endpoint,
+            oidc_token,
+        });
+
+        return client.runService(serviceName, file);
+    }
+
+    /**
+     * Create new service in OSCAR
+     * @param oscar_endpoint oscare endpoint
+     * @param service service object to create
+     * @returns service created
+     */
+    createService(oscar_endpoint: string, service: Service) {
+        const oidc_token = this.getAccessToken();
+        const client: Client = new Client({
+            clusterId: '1',
+            oscar_endpoint,
+            oidc_token,
+        });
+
+        const request: any = {
+            ...service,
+        };
+        return client.createService(request);
     }
 }