+++ src/environments/endpoints.ts
@@ -11,6 +11,11 @@ export const endpoints = {
     toolsSummary: '/catalog/tools/detail',
     tool: '/catalog/tools/:name/metadata',
     toolConfiguration: '/catalog/tools/:name/config',
+    secrets: '/secrets',
     userStats: '/deployments/stats/user',
     clusterStats: '/deployments/stats/cluster',
+    //OSCAR endpoints
+    services: '/services',
+    serviceByName: '/services/:name',
+    runService: '/run/:name',
 };